+++ src/main/java/creational/builder/IBuilder.java
@@ -1,10 +1,10 @@
-package creational.Builder;
+package creational.builder;
 
-import creational.Builder.enums.CarType;
-import creational.Builder.enums.Transmission;
-import creational.Builder.models.Engine;
-import creational.Builder.models.GPSNavigator;
-import creational.Builder.models.TripComputer;
+import creational.builder.enums.CarType;
+import creational.builder.enums.Transmission;
+import creational.builder.models.Engine;
+import creational.builder.models.GPSNavigator;
+import creational.builder.models.TripComputer;
 
 public interface IBuilder {
     void setCarType(CarType type);
