+++ src/main/java/creational/builder/enums/CarType.java
@@ -1,4 +1,4 @@
-package creational.Builder.enums;
+package creational.builder.enums;
 
 public enum CarType {
     CITY_CAR, SPORTS_CAR, SUV
