+++ src/main/java/creational/builder/enums/Transmission.java
@@ -1,4 +1,4 @@
-package creational.Builder.enums;
+package creational.builder.enums;
 
 public enum Transmission {
     SINGLE_SPEED, MANUAL, AUTOMATIC, SEMI_AUTOMATIC
