+++ src/main/java/creational/builder/models/Engine.java
@@ -1,4 +1,4 @@
-package creational.Builder.models;
+package creational.builder.models;
 
 public class Engine {
     private final double volume;
