+++ src/main/java/creational/factory_method/Button.java
@@ -1,4 +1,4 @@
-package creational.FactoryMethod;
+package creational.factory_method;
 
 public interface Button {
     void render();
