+++ src/main/java/creational/prototype/BundleShapeCache.java
@@ -1,4 +1,4 @@
-package creational.Prototype;
+package creational.prototype;
 
 import java.util.HashMap;
 import java.util.Map;
