+++ src/test/java/creational/SingletonTest.java
@@ -1,6 +1,6 @@
 package creational;
 
-import creational.Singleton.Singleton;
+import creational.singleton.Singleton;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
